ci: handle tests
